requirements.txt
--extra-index-url=https://quanttide-pypi.pkg.coding.net/qtopen-django/drf-remote-auth/simple
drf-remote-auth==0.1.0
--extra-index-url=https://quanttide-pypi.pkg.coding.net/qtopen-django/drf-admin/simple
drf-admin==0.2.0
django-reversion==5.0.1